# bibliotheque-backend/src/main/java/com/ibizabroker/bibliotheque/utilisateurs/web/UserCreateRequest.java
package com.ibizabroker.bibliotheque.utilisateurs.web;

import com.ibizabroker.bibliotheque.shared.util.Canonical;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Size;
import java.util.Set;

public class UserCreateRequest {

    @NotBlank(message = "Le nom d'utilisateur est obligatoire")
    private String username;

    @NotBlank(message = "Le nom est obligatoire")
    private String name;

    @Size(min = 6, message = "Le mot de passe doit contenir au moins 6 caractères")
    private String password;

    private Set<String> roles;

    public UserCreateRequest() {}

    // Forme canonique dès la réception : doublons et recherches ignorent la casse et les espaces.
    public String getUsername() { return username; }
    public void setUsername(String username) { this.username = Canonical.username(username); }

    public String getName() { return name; }
    public void setName(String name) { this.name = Canonical.texte(name); }

    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }

    public Set<String> getRoles() { return roles; }
    public void setRoles(Set<String> roles) { this.roles = roles; }
}