# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "bibliothèque",
    "description" : "API de gestion de bibliothèque. Pour tester : bouton « Authorize », section « keycloak », saisir un compte (ex. admin / admin123 ou A1 / A1123), laisser client_id = bibliotheque-frontend, sans client_secret. Pour voir l'access token et le refresh token : POST /auth/token (à relancer quand l'access token expire)."
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "keycloak" : [ ]
  }, {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/api/v1/users/{id}" : {
      "get" : {
        "tags" : [ "Utilisateurs" ],
        "summary" : "Obtenir un utilisateur par son identifiant",
        "operationId" : "getUserById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Utilisateurs" ],
        "summary" : "Modifier un utilisateur existant",
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users" : {
      "get" : {
        "tags" : [ "Utilisateurs" ],
        "summary" : "Lister les utilisateurs (pagination)",
        "operationId" : "getAllUsers",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "userId"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageUserResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Utilisateurs" ],
        "summary" : "Créer un nouvel utilisateur",
        "operationId" : "addUserByAdmin",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reservations" : {
      "get" : {
        "tags" : [ "Réservations" ],
        "summary" : "Lister les réservations (un ADHERENT ne voit que les siennes, RS-05)",
        "operationId" : "lister",
        "parameters" : [ {
          "name" : "statut",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "EN_ATTENTE", "DISPONIBLE", "ANNULEE", "EXPIREE", "HONOREE" ]
          }
        }, {
          "name" : "adherentId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Liste des réservations",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReservationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Réservations" ],
        "summary" : "Créer une réservation (un ADHERENT seulement pour lui-même, RS-04)",
        "operationId" : "creer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReservationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "livreId ou adherentId manquant",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "201" : {
            "description" : "Réservation créée",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Livre ou adhérent introuvable",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Règle de gestion violée (RG-01, RG-02 ou RG-03)",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Action non autorisée pour ce rôle",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reservations/{id}/annuler" : {
      "patch" : {
        "tags" : [ "Réservations" ],
        "summary" : "Annuler une réservation (un ADHERENT uniquement la sienne, RS-03)",
        "operationId" : "annuler",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Réservation introuvable",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Statut ne permettant pas l'annulation (RG-05 / RG-06)",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Réservation annulée",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Réservation d'un autre adhérent",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reservations/{id}" : {
      "get" : {
        "tags" : [ "Réservations" ],
        "summary" : "Consulter une réservation (un ADHERENT uniquement la sienne, RS-03)",
        "operationId" : "consulter",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Réservation introuvable",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Réservation trouvée",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Réservation d'un autre adhérent",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Réservations" ],
        "summary" : "Supprimer une réservation (bibliothécaire uniquement)",
        "operationId" : "supprimer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Réservation supprimée"
          },
          "404" : {
            "description" : "Réservation introuvable"
          },
          "403" : {
            "description" : "Action réservée au bibliothécaire (RS-02)"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UserCreateRequest" : {
        "required" : [ "name", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "password" : {
            "maxLength" : 2147483647,
            "minLength" : 6,
            "type" : "string"
          },
          "roles" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "ReservationRequest" : {
        "type" : "object",
        "properties" : {
          "livreId" : {
            "type" : "integer",
            "description" : "Id du livre à réserver (bookId)",
            "format" : "int32",
            "example" : 1
          },
          "adherentId" : {
            "type" : "integer",
            "description" : "Id de l'adhérent qui réserve (userId)",
            "format" : "int32",
            "example" : 2
          }
        }
      },
      "ReservationResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "livreId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "livreNom" : {
            "type" : "string"
          },
          "adherentId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "adherentNom" : {
            "type" : "string"
          },
          "dateReservation" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateExpiration" : {
            "type" : "string",
            "format" : "date-time"
          },
          "statut" : {
            "type" : "string",
            "enum" : [ "EN_ATTENTE", "DISPONIBLE", "ANNULEE", "EXPIREE", "HONOREE" ]
          }
        }
      },
...
      "PageUserResponse" : {
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/Sort"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/Pageable"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sort" : {
            "$ref" : "#/components/schemas/Sort"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paged" : {
            "type" : "boolean"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "Sort" : {
        "type" : "object",
        "properties" : {
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "keycloak" : {
        "type" : "oauth2",
        "description" : "Connexion Keycloak (realm bibliotheque)",
        "flows" : {
          "password" : {
            "tokenUrl" : "http://localhost:9999/realms/bibliotheque/protocol/openid-connect/token",
            "scopes" : { }
          }
        }
      },
      "bearerAuth" : {
        "type" : "http",
        "description" : "Jeton d'accès Keycloak déjà obtenu (sans le préfixe Bearer)",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}